+++ app/[locale]/search/Categories.tsx
